requirements.txt
django-sass
django-fontawesome-5